move (human) moderation front-end into election_snooper app
