+++ every_election/apps/election_snooper/templates/election_snooper/sub_menu.html
@@ -1,30 +1,30 @@
 <ul class="menu">
   <li
-    {% if 'election_radar' in request.path and not request.GET.status and not request.GET.pk %}
+    {% if 'moderation_queue' not in request.path and not request.GET.status and not request.GET.pk %}
       class="active"
     {% endif %} >
     <a href="{% url "snooped_election_view" %}">All</a>
   </li>
   <li
-    {% if 'election_radar' in request.path and request.GET.status == 'new' %}
+    {% if 'moderation_queue' not in request.path and request.GET.status == 'new' %}
       class="active"
     {% endif %} >
     <a href="{% url "snooped_election_view" %}?status=new">New</a>
   </li>
   <li
-    {% if 'election_radar' in request.path and request.GET.status == 'election' %}
+    {% if 'moderation_queue' not in request.path and request.GET.status == 'election' %}
       class="active"
     {% endif %} >
     <a href="{% url "snooped_election_view" %}?status=election">Elections</a>
   </li>
   <li
-    {% if 'election_radar' in request.path and request.GET.status == 'out_of_scope' %}
+    {% if 'moderation_queue' not in request.path and request.GET.status == 'out_of_scope' %}
       class="active"
     {% endif %} >
     <a href="{% url "snooped_election_view" %}?status=out_of_scope">Out of scope</a>
   </li>
   <li
-    {% if 'election_radar' in request.path and request.GET.status == 'duplicate' %}
+    {% if 'moderation_queue' not in request.path and request.GET.status == 'duplicate' %}
       class="active"
     {% endif %} >
     <a href="{% url "snooped_election_view" %}?status=duplicate">Duplicate</a>
