+++ every_election/apps/election_snooper/urls.py
@@ -3,7 +3,14 @@ from election_snooper import views
 
 
 urlpatterns = [
-    url(r'^$',
+    url(
+        r'^$',
         views.SnoopedElectionView.as_view(),
-        name='snooped_election_view'),
+        name='snooped_election_view',
+    ),
+    url(
+        r'^moderation_queue/$',
+        views.ModerationQueueView.as_view(),
+        name='election_moderation_queue',
+    ),
 ]
