+++ every_election/apps/election_snooper/views/__init__.py
@@ -1,0 +1,2 @@
+from .bot_review import *  # noqa
+from .human_review import *  # noqa
