+++ every_election/apps/elections/forms.py
@@ -4,13 +4,7 @@ from django.db import models
 from organisations.models import Organisation
 from organisations.models import OrganisationDivisionSet
 
-from .models import (
-    ElectionSubType,
-    ElectionType,
-    ModerationHistory,
-    ModerationStatus,
-    ModerationStatuses,
-)
+from .models import ElectionSubType, ElectionType
 
 from dc_theme import forms as dc_forms
 
@@ -182,24 +176,3 @@ class NoticeOfElectionForm(forms.Form):
         http://doncaster.gov.uk/services/the-council-democracy/notice-of-elections
         """
         return document
-
-
-class ModerationHistoryForm(forms.ModelForm):
-
-    class Meta:
-        model = ModerationHistory
-        fields = ['status', ]
-        widgets = {
-            'status': forms.widgets.RadioSelect()
-        }
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.fields['status'].queryset = ModerationStatus\
-            .objects\
-            .filter(short_label__in=[
-                ModerationStatuses.approved.value,
-                ModerationStatuses.suggested.value,
-                ModerationStatuses.rejected.value,
-            ])
-        self.fields['status'].empty_label=None
