+++ every_election/apps/elections/views/__init__.py
@@ -1,3 +1,2 @@
 from .id_creator import *  # noqa
-from .moderation_queue import *  # noqa
 from .general import *  # noqa
